requirements.txt
alabaster==0.7.12
amqp==5.0.9
apturl==0.5.2
asgiref==3.4.1
attrs==19.3.0
Babel==2.9.1
bcrypt==3.1.7
beautifulsoup4==4.9.3
billiard==3.6.4.0
blinker==1.4
braintree==4.11.0
Brlapi==0.7.0
celery==5.2.3
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.3
Click==7.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.26.0
colorama==0.4.3
command-not-found==0.3
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
cryptography==3.4.7
cupshelpers==1.0
cycler==0.10.0
dateutils==0.6.12
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.5
django-appconf==1.0.4
django-bootstrap3==15.0.0
django-bootstrap4==3.0.1
django-cors-headers==3.7.0
django-crispy-forms==1.9.2
django-decouple==2.1
django-extensions==3.1.5
django-fontawesome-5==1.0.18
django-heroku==0.3.1
django-imagekit==4.0.2
django-material==1.9.0
django-registration==3.2
django-rest-email-manager==1.1.0
django-rest-framework-braces==0.3.4
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.1
django-tinymce==3.3.0
django-url-or-relative-url-field==0.1.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
docutils==0.17.1
drf-yasg==1.17.1
duplicity==0.8.12.0
entrypoints==0.3
fancycompleter==0.9.1
fasteners==0.14.1
filelock==3.4.2
flake8==4.0.1
future==0.18.2
GDAL==3.0.4
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.14.0
idna==3.2
imagesize==1.3.0
importanize==0.7.0
importlib-metadata==4.11.1
inflection==0.5.0
ipython-genutils==0.2.0
itypes==1.2.0
Jinja2==2.10.1
jsonschema==3.2.0
jupyter-core==4.6.3
keyring==18.0.1
kiwisolver==1.0.1
kombu==5.2.3
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
livereload==2.6.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.1.2
mccabe==0.6.1
mock==4.0.3
monotonic==1.5
more-itertools==4.2.0
nbformat==5.0.4
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
OWSLib==0.19.1
packaging==20.4
paramiko==2.6.0
pathlib2==2.3.7.post1
pbr==5.8.0
pdbpp==0.10.3
pexpect==4.6.0
pilkit==2.0
Pillow==8.3.1
pipenv==2022.1.8
platformdirs==2.4.1
plotly==4.4.1
pluggy==1.0.0
prompt-toolkit==3.0.26
protobuf==3.6.1
psycopg2==2.9.1
py==1.11.0
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.20
pycups==1.9.73
pyflakes==2.4.0
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyproj==2.5.0
PyQt5==5.14.1
pyrepl==0.9.0
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.1
python-dotenv==0.13.0
python-http-client==3.2.7
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.26.0
requests-unixsocket==0.2.0
retrying==1.3.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
SecretStorage==2.3.1
sendgrid==6.3.1
simplejson==3.16.0
sip==4.19.21
six==1.16.0
snowballstemmer==2.2.0
soupsieve==1.9.4
Sphinx==4.4.0
sphinx-autobuild==2021.3.14
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
stripe==1.82.1
systemd-python==234
testresources==2.0.1
toml==0.10.2
tornado==6.1
tox==3.24.5
traitlets==4.3.3
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.6
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.13.0
virtualenv-clone==0.5.7
wadllib==1.3.3
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
whitenoise==4.1.4
wmctrl==0.4
xkit==0.0.0
zipp==1.0.0
